Move state-forwarding functions for the API client to the api module in mullvad-daemon
